+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ In development
 
 - add get_quantity on an account to retrieve the balance in quantity of the accound (for non currency based accounts)
 - fix bug when providing a float instead of a Decimal to a numeric value (@gregorias)
+- support new format for date for 2.7/2.8 (@MisterY, @sdementen)
 
 Version 0.13.0 (2017-10-08)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~