+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ In development
 - fix bug when providing a float instead of a Decimal to a numeric value (@gregorias)
 - support new format for date for 2.7/2.8 (@MisterY, @sdementen)
 - fix bug where transactions based on deleted scheduled transactions cause exceptions (@spookylukey)
+- fix bug (#58) where large Decimals where raising an sql exception instead of a ValueError exception (@sdementen)
 
 Version 0.13.0 (2017-10-08)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~