+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ In development
 - support new format for date for 2.7/2.8 (@MisterY, @sdementen)
 - fix bug where transactions based on deleted scheduled transactions cause exceptions (@spookylukey)
 - fix bug (#58) where large Decimals where raising an sql exception instead of a ValueError exception (@sdementen)
+- add Recurrence to global imports + add documentation to Recurrence (@MisterY)
 
 Version 0.13.0 (2017-10-08)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~